refactor: NASA-00 update about page styles
