+++ src/domain/about/modalAbout/_modalAbout.module.scss
@@ -1,6 +1,23 @@
 @import '../../../assets/styles/mixins';
 @import '../../../assets/styles/variables';
 
+.overlay {
+  content: '';
+
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  left: 0;
+
+  display: block;
+
+  width: 100vw;
+  height: 100vh;
+
+  opacity: 0.9;
+  background-color: #161720;
+}
+
 .modal {
   position: fixed;
   z-index: 2;
@@ -13,12 +30,11 @@
   width: 40%;
   height: fit-content;
   margin: auto;
-  padding: 2em;
+  padding: 4rem;
 
   background-color: #1f212d;
-  border: none;
+  border: 1px solid $border-color-default;
   border-radius: 24px;
-  box-shadow: 0 0 10px 10px rgb(1 1 1 / 46.3%);
 
   &::backdrop {
     background: rgb(0 0 0);
@@ -29,8 +45,8 @@
 
     position: absolute;
     z-index: 2;
-    top: 1.5em;
-    right: 1.5em;
+    top: 2.25rem;
+    right: 2.25rem;
 
     padding: 0;
