+++ src/pages/aboutPage/_about.module.scss
@@ -3,9 +3,7 @@
 
 .about {
   flex-direction: column;
-  margin: 2rem auto;
-
-  // outline: 4px solid $background-color-secondary;
+  margin: 6rem auto;
 
   .title,
   .title2 {
@@ -29,7 +27,9 @@
     }
 
     .horizontal-line {
-      width: 45%;
+      display: block;
+      flex-grow: 1;
+
       height: 1px;
 
       color: $background-color-secondary;
@@ -72,12 +72,45 @@
     width: 920px;
     height: 950px;
     margin: 7rem auto;
-
-    border: 1px solid #dcebea;
+    background: linear-gradient(
+      180deg,
+      rgb(220 235 234 / 100%) 0%,
+      rgb(220 235 234 / 77.5%) 6%,
+      rgb(220 235 234 / 35.8%) 13%,
+      rgb(220 235 234 / 0%) 32%,
+      rgb(220 235 234 / 0%) 50%,
+      rgb(220 235 234 / 0%) 68%,
+      rgb(220 235 234 / 35.5%) 84%,
+      rgb(220 235 234 / 66.3%) 92%,
+      rgb(220 235 234 / 100%) 100%
+    );
+    border: 1px solid $border-color-default;
     border-radius: 50%;
+
+    &::after {
+      content: '';
+
+      position: absolute;
+      z-index: 1;
+      top: 0;
+      left: 0;
+
+      display: block;
+
+      width: 100%;
+      height: 100%;
+
+      background-color: #161720;
+      border-radius: 50%;
+
+      @media (width <= 850px) {
+        display: none;
+      }
+    }
 
     .dot {
       position: absolute;
+      z-index: 2;
 
       width: 10px;
       height: 10px;
@@ -89,6 +122,8 @@
       animation: flash-dot 2s linear infinite;
 
       .dot-label {
+        @include font-arimo-bold;
+
         position: absolute;
 
         width: 200px;
@@ -165,6 +200,7 @@
 
     .card-blocks {
       position: relative;
+      z-index: 2;
       top: 13%;
 
       display: flex;
@@ -190,6 +226,7 @@
       cursor: pointer;
 
       position: absolute;
+      z-index: 2;
 
       width: 80px;
       height: 80px;
@@ -274,11 +311,11 @@
     max-width: 1200px;
     height: fit-content;
     margin: 0 auto;
-    padding: 2rem;
+    padding: 4rem;
 
     background-color: #1f212d;
-    border-radius: 20px;
-    box-shadow: 0 0 10px 10px rgb(39 39 39 / 46.3%);
+    border: 1px solid $border-color-default;
+    border-radius: 24px;
 
     .paragraph-block {
       position: relative;
@@ -287,11 +324,12 @@
       justify-content: space-between;
 
       .paragraph {
-        max-width: 50%;
+        max-width: 45%;
 
         .paragraph-title {
-          margin: 1rem;
-          font-size: 21px;
+          margin-bottom: 1rem;
+          font-size: 18px;
+          opacity: 0.8;
 
           @media (width <= 950px) {
             font-size: 1rem;
@@ -303,9 +341,8 @@
         }
 
         .paragraph-desc {
-          margin: 1rem;
           font-size: 16px;
-          text-align: justify;
+          opacity: 0.8;
 
           @media (width <= 950px) {
             font-size: 1rem;
@@ -323,7 +360,7 @@
 
           width: 80px;
           height: 40px;
-          margin: 1rem;
+          margin-top: 1.5rem;
 
           border-radius: 10px;
